fix(deepagents-code): report onboard upstream provider in Deep Code TUI
Managed inference always routes through the OpenAI-compatible adapter, so
the Deep Code TUI status bar and the model-identity system prompt reported
the wire provider "openai" instead of the onboard-selected upstream. Show
the upstream provider (from NEMOCLAW_UPSTREAM_PROVIDER) at those surfaces
while the adapter key stays "openai" for routing.
