+++ agents/langchain-deepagents-code/managed-dcode-runtime.py
@@ -37,6 +37,8 @@ _CREDENTIAL_ENV_NAMES = {
     "OTEL_EXPORTER_OTLP_TRACES_HEADERS",
 }
 _OPENSHELL_ENV_PLACEHOLDER_PREFIX = "openshell:resolve:env:"
+_UPSTREAM_PROVIDER_ENV = "NEMOCLAW_UPSTREAM_PROVIDER"
+_DISPLAY_PROVIDER_NAME = re.compile(r"[A-Za-z0-9][A-Za-z0-9._-]{0,63}")
 _MCP_SERVER_NAME = re.compile(r"[A-Za-z][A-Za-z0-9_-]{0,63}")
 _MCP_ENV_NAME = re.compile(r"[A-Za-z_][A-Za-z0-9_]{0,127}")
 _MCP_DNS_NAME = re.compile(
@@ -856,6 +858,24 @@ def managed_inference_base_url() -> str:
     ):
         raise RuntimeError("managed inference base URL is invalid")
     return value
+
+
+def managed_display_provider(adapter_provider: object) -> str:
+    """Return the provider label to show for the managed inference adapter.
+
+    Managed inference always routes through the OpenAI-compatible adapter, so
+    Deep Agents Code reports the wire provider (`openai`) in the status bar and
+    the model-identity system prompt. Substitute the onboard-selected upstream
+    provider so those surfaces match the launch page. Falls back to the
+    unchanged adapter provider when no valid upstream is present.
+    """
+    adapter = adapter_provider if isinstance(adapter_provider, str) else ""
+    upstream = os.environ.get(_UPSTREAM_PROVIDER_ENV, "").strip()
+    if not upstream or upstream == adapter:
+        return adapter
+    if _DISPLAY_PROVIDER_NAME.fullmatch(upstream) is None:
+        return adapter
+    return upstream
 
 
 def assert_safe_runtime() -> None:
